separate model initialization and model state
