+++ app/assets/javascripts/admin/admin_rest_api.js
@@ -31,11 +31,13 @@ import type {
   APITracingType,
   APIFeatureToggles,
   APIOrganizationType,
+  ServerTracingType,
 } from "admin/api_flow_types";
 import type { QueryObjectType } from "admin/task/task_search_form";
 import type { NewTaskType, TaskCreationResponseType } from "admin/task/task_create_bulk_view";
-import type { DatasetConfigurationType } from "oxalis/store";
-import type { ServerTracingType } from "oxalis/model";
+import type {
+  DatasetConfigurationType,
+} from "oxalis/store";
 
 const MAX_SERVER_ITEMS_PER_RESPONSE = 1000;
