+++ app/assets/javascripts/oxalis/geometries/materials/plane_material_factory.js
@@ -27,9 +27,18 @@ import { calculateGlobalPos } from "oxalis/controller/viewmodes/plane_controller
 import { getActiveCellId, getVolumeTool } from "oxalis/model/accessors/volumetracing_accessor";
 import getMainFragmentShader from "oxalis/shaders/main_data_fragment.glsl";
 import { getPackingDegree } from "oxalis/model/bucket_data_handling/data_rendering_logic";
-import { getResolutions, isRgb, getByteCount } from "oxalis/model/accessors/dataset_accessor.js";
+import {
+  getColorLayers,
+  getResolutions,
+  isRgb,
+  getByteCount,
+} from "oxalis/model/accessors/dataset_accessor.js";
 
 const DEFAULT_COLOR = new THREE.Vector3([255, 255, 255]);
+
+function getColorLayerNames() {
+  return getColorLayers(Store.getState().dataset).map(layer => sanitizeName(layer.name));
+}
 
 class PlaneMaterialFactory extends AbstractPlaneMaterialFactory {
   planeID: OrthoViewType;
@@ -180,10 +189,7 @@ class PlaneMaterialFactory extends AbstractPlaneMaterialFactory {
     }
 
     // Add weight/color uniforms
-    const colorLayerNames = _.map(Model.getColorLayers(), colorLayer =>
-      sanitizeName(colorLayer.name),
-    );
-    for (const name of colorLayerNames) {
+    for (const name of getColorLayerNames()) {
       this.uniforms[`${name}_weight`] = {
         type: "f",
         value: 1,
@@ -363,7 +369,7 @@ class PlaneMaterialFactory extends AbstractPlaneMaterialFactory {
   }
 
   getFragmentShader(): string {
-    const colorLayerNames = _.map(Model.getColorLayers(), b => sanitizeName(b.name));
+    const colorLayerNames = getColorLayerNames();
     const segmentationLayer = Model.getSegmentationLayer();
     const segmentationName = sanitizeName(segmentationLayer ? segmentationLayer.name : "");
     const { dataset } = Store.getState();
