+++ app/assets/javascripts/oxalis/model/accessors/volumetracing_accessor.js
@@ -6,7 +6,7 @@ import Maybe from "data.maybe";
 import { getRequestLogZoomStep } from "oxalis/model/accessors/flycam_accessor";
 import type { TracingType, VolumeTracingType, OxalisState } from "oxalis/store";
 import type { VolumeToolType, ContourModeType } from "oxalis/constants";
-import type { ServerTracingType, ServerVolumeTracingType } from "oxalis/model";
+import type { ServerTracingType, ServerVolumeTracingType } from "admin/api_flow_types";
 
 export function getVolumeTracing(tracing: TracingType): Maybe<VolumeTracingType> {
   if (tracing.type === "volume") {
