+++ app/assets/javascripts/oxalis/model/reducers/skeletontracing_reducer.js
@@ -40,8 +40,8 @@ import type {
   TreeType,
 } from "oxalis/store";
 import DiffableMap from "libs/diffable_map";
-import type { ServerNodeType, ServerBranchPointType } from "oxalis/model";
 import type { ActionType } from "oxalis/model/actions/actions";
+import type {ServerNodeType, ServerBranchPointType} from "admin/api_flow_types";
 import Maybe from "data.maybe";
 import ErrorHandling from "libs/error_handling";
