+++ app/assets/javascripts/test/fixtures/volumetracing_server_objects.js
@@ -3,20 +3,17 @@ export const tracing = {
   dataLayer: {
     name: "64007765-cef9-4e31-b206-dba795b5be17",
     category: "segmentation",
-    boundingBox: {
-      topLeft: [3840, 4220, 2304],
-      width: 128,
-      height: 131,
-      depth: 384,
-    },
     resolutions: [[1, 1, 1]],
     fallbackLayer: "segmentation",
     elementClass: "uint16",
     largestSegmentId: 21890,
   },
+  elementClass: "uint16",
+  boundingBox: { topLeft: { x: 0, y: 0, z: 0 }, width: 1024, height: 1024, depth: 1024 },
   zoomLevel: 0,
   editPosition: [3904, 4282, 2496],
   editRotation: [0, 0, 0],
+  id: "segmentation",
 };
 
 export const annotation = {
@@ -29,7 +26,7 @@ export const annotation = {
   stats: null,
   restrictions: { allowAccess: true, allowUpdate: true, allowFinish: true, allowDownload: true },
   formattedHash: "f043e7",
-  content: { id: "47e37793-d0be-4240-a371-87ce68561a13", typ: "volume" },
+  content: { id: "segmentation", typ: "volume" },
   dataSetName: "ROI2017_wkw",
   dataStore: { name: "localhost", url: "http://localhost:9000", typ: "webknossos-store" },
   settings: {
